+++ src/components/internal/Loader.vue
@@ -5,8 +5,8 @@ withDefaults(defineProps<{
     backgroundColor?: string;
 }>(), {
     size: 150,
-    color: '#FF7E29',
-    backgroundColor: '#ff7f29a1',
+    color: 'rgb(var(--color-spartan-primary-600))',
+    backgroundColor: 'rgb(var(--color-spartan-primary-600))',
 });
 </script>
 
@@ -21,24 +21,6 @@ withDefaults(defineProps<{
                 repeatCount="indefinite"
                 begin="-0.9166666666666666s"
             ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
-                begin="-0.9166666666666666s"
-            ></animate>
         </circle>
         <circle cx="71.65063509461098" cy="62.5" :fill="backgroundColor" r="5">
             <animate
@@ -47,24 +29,6 @@ withDefaults(defineProps<{
                 times="0;0.1;0.2;0.3;1"
                 dur="1s"
                 repeatCount="indefinite"
-                begin="-0.8333333333333334s"
-            ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
                 begin="-0.8333333333333334s"
             ></animate>
         </circle>
@@ -77,24 +41,6 @@ withDefaults(defineProps<{
                 repeatCount="indefinite"
                 begin="-0.75s"
             ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
-                begin="-0.75s"
-            ></animate>
         </circle>
         <circle cx="50" cy="75" :fill="backgroundColor" r="5">
             <animate
@@ -103,24 +49,6 @@ withDefaults(defineProps<{
                 times="0;0.1;0.2;0.3;1"
                 dur="1s"
                 repeatCount="indefinite"
-                begin="-0.6666666666666666s"
-            ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
                 begin="-0.6666666666666666s"
             ></animate>
         </circle>
@@ -133,24 +61,6 @@ withDefaults(defineProps<{
                 repeatCount="indefinite"
                 begin="-0.5833333333333334s"
             ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
-                begin="-0.5833333333333334s"
-            ></animate>
         </circle>
         <circle cx="28.34936490538903" cy="62.5" :fill="backgroundColor" r="5">
             <animate
@@ -159,24 +69,6 @@ withDefaults(defineProps<{
                 times="0;0.1;0.2;0.3;1"
                 dur="1s"
                 repeatCount="indefinite"
-                begin="-0.5s"
-            ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
                 begin="-0.5s"
             ></animate>
         </circle>
@@ -189,24 +81,6 @@ withDefaults(defineProps<{
                 repeatCount="indefinite"
                 begin="-0.4166666666666667s"
             ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
-                begin="-0.4166666666666667s"
-            ></animate>
         </circle>
         <circle cx="28.34936490538903" cy="37.50000000000001" :fill="backgroundColor" r="5">
             <animate
@@ -215,24 +89,6 @@ withDefaults(defineProps<{
                 times="0;0.1;0.2;0.3;1"
                 dur="1s"
                 repeatCount="indefinite"
-                begin="-0.3333333333333333s"
-            ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
                 begin="-0.3333333333333333s"
             ></animate>
         </circle>
@@ -245,24 +101,6 @@ withDefaults(defineProps<{
                 repeatCount="indefinite"
                 begin="-0.25s"
             ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
-                begin="-0.25s"
-            ></animate>
         </circle>
         <circle cx="49.99999999999999" cy="25" :fill="backgroundColor" r="5">
             <animate
@@ -271,24 +109,6 @@ withDefaults(defineProps<{
                 times="0;0.1;0.2;0.3;1"
                 dur="1s"
                 repeatCount="indefinite"
-                begin="-0.16666666666666666s"
-            ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
                 begin="-0.16666666666666666s"
             ></animate>
         </circle>
@@ -301,24 +121,6 @@ withDefaults(defineProps<{
                 repeatCount="indefinite"
                 begin="-0.08333333333333333s"
             ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
-                begin="-0.08333333333333333s"
-            ></animate>
         </circle>
         <circle cx="71.65063509461096" cy="37.499999999999986" :fill="backgroundColor" r="5">
             <animate
@@ -327,24 +129,6 @@ withDefaults(defineProps<{
                 times="0;0.1;0.2;0.3;1"
                 dur="1s"
                 repeatCount="indefinite"
-                begin="0s"
-            ></animate>
-            <animate
-                attributeName="fill"
-                :values="
-                    backgroundColor +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    color +
-                    ';' +
-                    backgroundColor +
-                    ';' +
-                    backgroundColor
-                "
-                repeatCount="indefinite"
-                times="0;0.1;0.2;0.3;1"
-                dur="1s"
                 begin="0s"
             ></animate>
         </circle>